Kill child process properly

+++ http-servers/kill_server.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
 
+set -e
+
 pid=$(cat tmp/server.pid)
-kill -s KILL $pid
+kill -s TERM $pid
+
+sleep 1
+if ps -p $pid > /dev/null; then
+  kill -s KILL $pid
+fi
